simplify app.js
