+++ index.js
@@ -8,7 +8,7 @@ import App from './src/components/App';
 import {name as appName} from './app.json';
 
 // if developing in Xcode, uncomment this line:
-AppRegistry.registerComponent(appName, () => App);
+// AppRegistry.registerComponent(appName, () => App);
 
 // if developing in Expo, uncomment this line:
-// registerRootComponent(App);
+registerRootComponent(App);
